Поправила замечания
