+++ js/game/levels/levelArtist-view.js
@@ -73,7 +73,7 @@ export default class LevelArtistView extends AbstractView {
         e.preventDefault();
         if (item.checked) {
           if (i === rightAnswerIndex) {
-            this.onAnswer(getLevel(setNextLevel(this.game)) ? Result.NEXT : Result.WIN);
+            this.onAnswer(getLevel(setNextLevel(this.game), this.model.data) ? Result.NEXT : Result.WIN);
           } else {
             let result = Result.DIE;
             if (this.game.lives <= 0 || this.game.time <= 0) {
