+++ js/game/levels/levelGenre-view.js
@@ -112,7 +112,7 @@ export default class LevelGenreView extends AbstractView {
         });
 
         if (countRightAnswers === count) {
-          this.onAnswer(getLevel(setNextLevel(this.game)) ? Result.NEXT : Result.WIN);
+          this.onAnswer(getLevel(setNextLevel(this.game), this.model.data) ? Result.NEXT : Result.WIN);
         } else {
           let result = Result.DIE;
           if (this.game.lives <= 0 || this.game.time <= 0) {
